update
Fixed the calculations
Added option to use E on towers in laneclear
Added option to enable/disable E + R finisher Note: It does not keep in
mind about any shields or heal spells.
Added option to always cast R even if the E stacks are gonna kill him.
Target can get shields or heal so it's smart to keep that on.